+++ src/test/polymer/polymer2-mixin-scanner_test.ts
@@ -314,7 +314,10 @@ Polymer.TestMixin = Polymer.woohoo(function TestMixin(base) {
           {name: 'customInstanceFunction', params: [], return: undefined},
           {
             name: 'customInstanceFunctionWithJSDoc',
-            params: [], return: undefined,
+            params: [], return: {
+              desc: 'The number 5, always.',
+              type: 'Number',
+            },
           },
           {
             name: 'customInstanceFunctionWithParams',
@@ -322,7 +325,26 @@ Polymer.TestMixin = Polymer.woohoo(function TestMixin(base) {
           },
           {
             name: 'customInstanceFunctionWithParamsAndJSDoc',
-            params: [{name: 'a'}, {name: 'b'}, {name: 'c'}], return: undefined,
+            params: [
+              {
+                name: 'a',
+                type: 'Number',
+                description: 'The first argument',
+              },
+              {
+                name: 'b',
+                type: 'Number',
+              },
+              {
+                name: 'c',
+                type: 'Number',
+                description: 'The third argument',
+              }
+            ],
+            return: {
+              desc: 'The number 7, always.',
+              type: 'Number',
+            },
           },
           {
             name: 'customInstanceFunctionWithParamsAndPrivateJSDoc',