+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 * `generateAnalysis()` now includes PolymerBehavior information in `metadata.polymer.behaviors` collection.
+* Types and descriptions are now extracted from method @param and @returns jsdoc annotations.
 
 ## [2.0.0-alpha.34] - 2017-03-20